Move flattened_varnames_list to main package
This is required to use the utility in Julia versions that support extensions.
