+++ ext/PathfinderTuringExt.jl
@@ -7,6 +7,7 @@ if isdefined(Base, :get_extension)
     using Pathfinder: Pathfinder
     using Random: Random
     using Turing: Turing
+    import Pathfinder: flattened_varnames_list
 else  # using Requires
     using ..Accessors: Accessors
     using ..DynamicPPL: DynamicPPL
@@ -14,46 +15,12 @@ else  # using Requires
     using ..Pathfinder: Pathfinder
     using ..Random: Random
     using ..Turing: Turing
+    import ..Pathfinder: flattened_varnames_list
 end
 
 # utilities for working with Turing model parameter names using only the DynamicPPL API
 
-"""
-    flattened_varnames_list(model::DynamicPPL.Model) -> Vector{Symbol}
-
-Get a vector of varnames as `Symbol`s with one-to-one correspondence to the
-flattened parameter vector.
-
-```julia
-julia> @model function demo()
-           s ~ Dirac(1)
-           x = Matrix{Float64}(undef, 2, 4)
-           x[1, 1] ~ Dirac(2)
-           x[2, 1] ~ Dirac(3)
-           x[3] ~ Dirac(4)
-           y ~ Dirac(5)
-           x[4] ~ Dirac(6)
-           x[:, 3] ~ arraydist([Dirac(7), Dirac(8)])
-           x[[2, 1], 4] ~ arraydist([Dirac(9), Dirac(10)])
-           return s, x, y
-       end
-demo (generic function with 2 methods)
-
-julia> flattened_varnames_list(demo())
-10-element Vector{Symbol}:
- :s
- Symbol("x[1,1]")
- Symbol("x[2,1]")
- Symbol("x[3]")
- Symbol("x[4]")
- Symbol("x[:,3][1]")
- Symbol("x[:,3][2]")
- Symbol("x[[2, 1],4][1]")
- Symbol("x[[2, 1],4][2]")
- :y
-```
-"""
-function flattened_varnames_list(model::DynamicPPL.Model)
+function Pathfinder.flattened_varnames_list(model::DynamicPPL.Model)
     varnames_ranges = varnames_to_ranges(model)
     nsyms = maximum(maximum, values(varnames_ranges))
     syms = Vector{Symbol}(undef, nsyms)
