+++ src/integration/turing.jl
@@ -1,0 +1,41 @@
+"""
+    flattened_varnames_list(model::DynamicPPL.Model) -> Vector{Symbol}
+
+Get a vector of varnames as `Symbol`s with one-to-one correspondence to the
+flattened parameter vector.
+
+!!! note
+    This function is only available when Turing has been loaded.
+
+# Examples
+
+```julia
+julia> @model function demo()
+           s ~ Dirac(1)
+           x = Matrix{Float64}(undef, 2, 4)
+           x[1, 1] ~ Dirac(2)
+           x[2, 1] ~ Dirac(3)
+           x[3] ~ Dirac(4)
+           y ~ Dirac(5)
+           x[4] ~ Dirac(6)
+           x[:, 3] ~ arraydist([Dirac(7), Dirac(8)])
+           x[[2, 1], 4] ~ arraydist([Dirac(9), Dirac(10)])
+           return s, x, y
+       end
+demo (generic function with 2 methods)
+
+julia> flattened_varnames_list(demo())
+10-element Vector{Symbol}:
+ :s
+ Symbol("x[1,1]")
+ Symbol("x[2,1]")
+ Symbol("x[3]")
+ Symbol("x[4]")
+ Symbol("x[:,3][1]")
+ Symbol("x[:,3][2]")
+ Symbol("x[[2, 1],4][1]")
+ Symbol("x[[2, 1],4][2]")
+ :y
+```
+"""
+function flattened_varnames_list end
